drop legacy SKIP_TF_GET environment variable
This was implemented primarily as a workaround for
https://github.com/hashicorp/go-getter/issues/104

If it was going to be retained, it would need to be rewritten as a
module option. Given that the original issue has been resolved (should
ship soon in an upcoming Terraform release), we'll just drop the option
for now (anyone needing it can keep using runway <v1 until a fixed
version of Terraform is released).