cfnlint now invoked with runpy.run_module, warn when missing templates (#284)
using runpy.run_module works the same as using "python -m" to directly invoke a module without need to import it in the script itself. combine with the argv context manager to inject the correct cli args, this imitates running the cfn-lint cli directly removing the need to maintain support for different versions in our codebase.