fixed TypeError when stack template from AWS contains non-JSON data type
If the template uploaded to AWS contains a date that is not quoted, it will be returned as a datetime object which cannot be handled by `json.dumps`.

The json encoder being added supports datetime as well as another data type that can appear in some AWS responses, Decimal, since I had the snippet that handles both laying around.