fix non-idempotent Auth@Edge function deploys
pyrsa/pyjwt create scripts with a shebang specific to the build system.
Removing them, and setuptools remnants containing various hash values,
stops new functions from being built on different systems.

It's possible the `PYTHONDONTWRITEBYTECODE` option is safe to apply
unconditionally to all function builds by default, but I'm not 100%
certain of the implications of that so I've gated it to an explicit
option for now.

All of the new options have been intentionally left off the docs for
this first pass where we may revise the names.